fix(project): 병합 후 발생하는 문제 해결

+++ src/main/java/com/fastcampus/toyproject4_team4/repository/PlaceRepository.java
@@ -3,5 +3,4 @@ package com.fastcampus.toyproject4_team4.repository;
 import com.fastcampus.toyproject4_team4.entity.place.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaceRepository extends JpaRepository<Place, Integer> {
-}
+public interface PlaceRepository extends JpaRepository<Place, Long> {}
